feat(adminpanel): added success toast for delete selected and delete all
renamed delete to delete selected, no longer displaying the admin panel table if no users exist, no
longer displaying the current user inside the admin panel table and, added validation to the
deleteAdmins checkbox server-sided to ensure its false or true and nothing else.

+++ packages/server/src/controllers/users/deleteSelectedUsers/index.ts
@@ -1,9 +1,18 @@
+import { deleteSelectedUsersMessage } from '@my-template/shared';
 import User from '../../../models/User.model';
 import { Route } from '../../../utils/types';
 
 const deleteSelectedUsers: Route = async (req, res) => {
   try {
     const { userIds } = req.body;
+
+    if (userIds.length === 0) {
+      console.error(deleteSelectedUsersMessage);
+
+      res.status(400).send({ message: deleteSelectedUsersMessage });
+
+      return;
+    }
 
     await User.destroy({ where: { id: userIds } });
